Fixed conditional dependencies
Since setup.py is not run during the installation of the wheel, it would depend on the Python version used for creating the wheel whether futures and ipaddress are included in the list of dependencies in the wheel metadata. The proper way to address this is to add them in extras_require as presented here.